fix memory leak

+++ Framework/Parser/BMP.hpp
@@ -54,7 +54,7 @@ namespace My {
                 img.bitcount = 32;
                 auto byte_count = img.bitcount >> 3;
                 img.pitch = ((img.Width * byte_count) + 3) & ~3;
-                img.data_size = img.pitch * img.Height;
+                img.data_size = (size_t)img.pitch * img.Height;
                 img.data = new uint8_t[img.data_size];
 
                 if (img.bitcount < 24) {
@@ -63,8 +63,8 @@ namespace My {
                     const uint8_t* pSourceData = reinterpret_cast<const uint8_t*>(buf.GetData()) + pFileHeader->BitsOffset;
                     for (int32_t y = img.Height - 1; y >= 0; y--) {
                         for (uint32_t x = 0; x < img.Width; x++) {
-                            auto dst = reinterpret_cast<R8G8B8A8Unorm*>(reinterpret_cast<uint8_t*>(img.data) + img.pitch * (img.Height - y - 1) + x * byte_count);
-                            auto src = reinterpret_cast<const R8G8B8A8Unorm*>(pSourceData + img.pitch * y + x * byte_count);
+                            auto dst = reinterpret_cast<R8G8B8A8Unorm*>(reinterpret_cast<uint8_t*>(img.data) + (ptrdiff_t)img.pitch * ((ptrdiff_t)img.Height - y - 1) + (ptrdiff_t)x * byte_count);
+                            auto src = reinterpret_cast<const R8G8B8A8Unorm*>(pSourceData + (ptrdiff_t)img.pitch * y + (ptrdiff_t)x * byte_count);
                             dst->data[2] = src->data[0];
                             dst->data[1] = src->data[1];
                             dst->data[0] = src->data[2];
